Reorganize media display packages and classes
