+++ lyrebird/src/main/java/moe/lyrebird/view/screens/media/display/twitter/photo/TwitterPhotoScreenController.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package moe.lyrebird.view.screens.media.twitter.photo;
+package moe.lyrebird.view.screens.media.display.twitter.photo;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
